src/manage_users/requirements.txt
# pg8000==1.19.4  <-- old, not needed anymore
# psycopg2==2.8.6  <-- old version
psycopg2-binary==2.9.5  # <-- actual ver

google-cloud-pubsub==2.16.0  # ok
google-cloud-secret-manager==2.16.0  # ok